Wlan: Merge ext cap from supplicant if bssTransition bit is set
Ext cap from supplicant is taken in consideration while sending assoc
request only if internetworking bit is set and thus if the
internetworking bit is not set in the ext cap from supplicant it is
not merged with ext cap of the assoc req.

Thus bssTransition bit info in ext cap provided by supplicant is
not taken in consideration if internetworking bit is not set.

To fix this merge the ext cap from supplicant if internetworking
or bssTransition bits are set.

Change-Id: I81d5d8c417dd7adc75b9864b625a00d53657df6c
CRs-Fixed: 964383